+++ src/Controller/LumaController.php
@@ -2,10 +2,13 @@
 
 namespace Luma\Framework\Controller;
 
+use Luma\Framework\Luma;
+use Luma\Framework\Messages\FlashMessage;
 use Luma\HttpComponent\Response;
 use Luma\HttpComponent\StreamBuilder;
 use Latte\Bridges\Tracy\TracyExtension;
 use Latte\Engine;
+use Luma\SecurityComponent\Authentication\Interface\UserInterface;
 use Tracy\Debugger;
 
 class LumaController
@@ -13,12 +16,15 @@ class LumaController
     private Engine $templateEngine;
     private static string $templateDirectory;
     private static string $cacheDirectory;
+    private array $errors = [];
+    private ?UserInterface $currentUser;
 
     public function __construct()
     {
         $this->templateEngine = new Engine();
         $this->templateEngine->addExtension(new TracyExtension());
         $this->templateEngine->setTempDirectory(static::$cacheDirectory);
+        $this->currentUser = Luma::getLoggedInUser();
     }
 
     /**
@@ -40,16 +46,17 @@ class LumaController
      * @param string $data
      * @param string $contentType
      * @param int $statusCode
+     * @param array $responseHeaders
      *
      * @return Response
      */
-    protected function respond(string $data, string $contentType = 'text/html', int $statusCode = 200): Response
+    protected function respond(string $data, string $contentType = 'text/html', int $statusCode = 200, array $responseHeaders = []): Response
     {
-        $responseHeaders = [
-            'Content-Type' => $contentType,
-        ];
+        if (!isset($responseHeaders['Content-Type'])) {
+            $responseHeaders['Content-Type'] = $contentType;
+        }
 
-        if (!Debugger::isEnabled()) {
+        if (!Debugger::isEnabled() && !isset($responseHeaders['Content-Length'])) {
             $responseHeaders['Content-Length'] = strlen($data);
         }
 
@@ -74,7 +81,31 @@ class LumaController
             : sprintf('%s.latte', $templatePath);
         $templatePath = sprintf('%s/%s', static::$templateDirectory, $templatePath);
 
+        if (!isset($data['messages'])) {
+            $data['messages'] = $this->getFlashMessages();
+        }
+
+        if (!isset($data['errors'])) {
+            $data['errors'] = $this->getErrors();
+        }
+
+        if (!isset($data['user'])) {
+            $data['user'] = $this->getLoggedInUser();
+        }
+
         return $this->respond($this->templateEngine->renderToString($templatePath, $data));
+    }
+
+    /**
+     * @param string $path
+     *
+     * @return Response
+     */
+    protected function redirect(string $path): Response
+    {
+        return $this->respond('', 'text/html', 302, [
+            'Location' => $path,
+        ]);
     }
 
     /**
@@ -85,5 +116,54 @@ class LumaController
     protected function json(array|object $data): Response
     {
         return $this->respond(json_encode($data), 'application/json');
+    }
+
+    /**
+     * @param string $error
+     *
+     * @return void
+     */
+    protected function addError(string $error): void
+    {
+        $this->errors[] = $error;
+    }
+
+    /**
+     * @return array
+     */
+    protected function getErrors(): array
+    {
+        return $this->errors;
+    }
+
+    /**
+     * @return UserInterface|null
+     */
+    protected function getLoggedInUser(): ?UserInterface
+    {
+        return $this->currentUser;
+    }
+
+    /**
+     * @param FlashMessage $message
+     * @param string $type
+     *
+     * @return void
+     */
+    protected function addFlashMessage(FlashMessage $message, string $type = 'info'): void
+    {
+        $_SESSION['messages'][$type][] = $message;
+    }
+
+    /**
+     * @return array
+     */
+    protected function getFlashMessages(): array
+    {
+        $messages = $_SESSION['messages'] ?? [];
+
+        unset($_SESSION['messages']);
+
+        return $messages;
     }
 }